Improve Vulkan instance building
Improved the Vulkan instance builder to support instance extension requiredments. In order to implement this, a logging module, diag, was added to aid in logging and recording program diagnositics.

In addition, an oversight that involved not copying the queryed physical devices to the interal array inside the PhysicalDeviceSelector constructor was fixed.